requirements/prod.txt
certifi==2020.6.20
influxdb==5.3.0
numpy==1.19.0
paho-mqtt==1.5.0
pandas==1.0.5
requests==2.24.0
